@@ -1,14 +1,9 @@
 package com.hands.on.controller;
-
 import com.hands.on.model.Person;
 import com.hands.on.repository.PersonRepository;
-import com.hands.on.stream.Topics;
+import com.hands.on.stream.PersonProducer;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.http.ResponseEntity;
-import org.springframework.messaging.Message;
-import org.springframework.messaging.MessageChannel;
-import org.springframework.messaging.support.MessageBuilder;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.HashMap;
@@ -18,39 +13,36 @@ import java.util.Map;
 @RequestMapping("/api")
 public class ApiController {
 
-    private MessageChannel messageChannel;
+    private PersonProducer personProducer;
     private PersonRepository repository;
 
     @Autowired
-    public ApiController(@Qualifier(Topics.INPUT) MessageChannel messageChannel,
+    public ApiController(PersonProducer personProducer,
                          PersonRepository repository
     ) {
-        this.messageChannel = messageChannel;
+        this.personProducer = personProducer;
         this.repository = repository;
     }
 
     @RequestMapping(path = "/person", method = RequestMethod.POST)
     public ResponseEntity<?> insert(@RequestBody Person person) {
-        Message<Person> message =  MessageBuilder.withPayload(person).build();
-        boolean isSuccess = this.messageChannel.send(message);
-        if(isSuccess) {
-            return ResponseEntity.noContent().build();
-        }
-        else {
+        boolean isSuccess = personProducer.produce(person);
+        if(!isSuccess) {
             Map json = new HashMap<>();
             json.put("errorMessage", "fail to publish Person in topic");
-            return ResponseEntity.unprocessableEntity().body(json);
+            return ResponseEntity.internalServerError().body(json);
         }
+        return ResponseEntity.noContent().build();
     }
 
     @RequestMapping(path = "/person", method = RequestMethod.GET)
     public ResponseEntity<?> get(@RequestParam(required = true) String email) {
-         Person person = this.repository.findByEmail(email);
-         if(person != null) {
-             return ResponseEntity.ok(person);
-         }
-         else {
-             return ResponseEntity.notFound().build();
-         }
+        Person person = this.repository.findByEmail(email);
+        if(person != null) {
+            return ResponseEntity.ok(person);
+        }
+        else {
+            return ResponseEntity.notFound().build();
+        }
     }
 }
